+++ hata/discord/client/request_helpers.py
@@ -22,7 +22,7 @@ from ..embed import EmbedBase
 from ..emoji import Emoji, parse_reaction
 from ..guild import Guild, GuildDiscovery
 from ..interaction import ApplicationCommand
-from ..message import Attachment, Message, MessageReference, MessageRepr
+from ..message import Attachment, Message, MessageRepr
 from ..oauth2 import Achievement
 from ..role import Role
 from ..scheduled_event import ScheduledEvent
@@ -128,7 +128,7 @@ def validate_message_to_delete(message):
     
     Parameters
     ----------
-    message : ``Message``, ``MessageReference``, ``MessageRepr``, `tuple` (`int`, `int`)
+    message : ``Message``, ``MessageRepr``, `tuple` (`int`, `int`)
         The message to validate for deletion.
     
     Returns
@@ -143,7 +143,7 @@ def validate_message_to_delete(message):
     Raises
     ------
     TypeError
-        If message was not given neither as ``Message``, ``MessageReference``, ``MessageRepr``, neither as
+        If message was not given neither as ``Message``, ``MessageRepr``, neither as
         `tuple` (`int`, `int`).
     """
     if isinstance(message, Message):
@@ -153,14 +153,11 @@ def validate_message_to_delete(message):
         if isinstance(message, MessageRepr):
             channel_id = message.channel_id
             message_id = message.id
-        elif isinstance(message, MessageReference):
-            channel_id = message.channel_id
-            message_id = message.message_id
         else:
             snowflake_pair = maybe_snowflake_pair(message)
             if snowflake_pair is None:
                 raise TypeError(
-                    f'`message` can be `{Message.__name__}`, `{MessageRepr.__name__}`, `{MessageReference.__name__}`, '
+                    f'`message` can be `{Message.__name__}`, `{MessageRepr.__name__}`, '
                     f'`tuple` of (`int`, `int`), got {message.__class__.__name__}; {message!r}.'
                 )
             
@@ -1264,7 +1261,7 @@ def get_channel_id_and_message_id(message):
     
     Parameters
     ----------
-    message : ``Message``, ``MessageRepr``, ``MessageReference``, `tuple` (`int`, `int`)
+    message : ``Message``, ``MessageRepr``, `tuple` (`int`, `int`)
         The message or it' representation.
     
     Returns
@@ -1281,7 +1278,6 @@ def get_channel_id_and_message_id(message):
     """
     # 1.: Message
     # 2.: MessageRepr
-    # 3.: MessageReference
     # 4.: None -> raise
     # 5.: `tuple` (`int`, `int`)
     # 6.: raise
@@ -1291,9 +1287,6 @@ def get_channel_id_and_message_id(message):
     elif isinstance(message, MessageRepr):
         channel_id = message.channel_id
         message_id = message.id
-    elif isinstance(message, MessageReference):
-        channel_id = message.channel_id
-        message_id = message.message_id
     elif message is None:
         raise TypeError(
             '`message` was given as `None`. Make sure to call message create methods with non-empty content(s).'
@@ -1302,7 +1295,7 @@ def get_channel_id_and_message_id(message):
         snowflake_pair = maybe_snowflake_pair(message)
         if snowflake_pair is None:
             raise TypeError(
-                f'`message` can be `{Message.__name__}`, `{MessageRepr.__name__}`, `{MessageReference.__name__}`, '
+                f'`message` can be `{Message.__name__}`, `{MessageRepr.__name__}`, '
                 f'`tuple` of (`int`, `int`), got {message.__class__.__name__}; {message!r}.'
             )
         
@@ -1317,7 +1310,7 @@ def get_message_and_channel_id_and_message_id(message):
     
     Parameters
     ----------
-    message : ``Message``, ``MessageRepr``, ``MessageReference``, `tuple` (`int`, `int`)
+    message : ``Message``, ``MessageRepr``, `tuple` (`int`, `int`)
         The message or it' representation.
     
     Returns
@@ -1336,7 +1329,6 @@ def get_message_and_channel_id_and_message_id(message):
     """
     # 1.: Message
     # 2.: MessageRepr
-    # 3.: MessageReference
     # 4.: None -> raise
     # 5.: `tuple` (`int`, `int`)
     # 6.: raise
@@ -1347,10 +1339,6 @@ def get_message_and_channel_id_and_message_id(message):
         channel_id = message.channel_id
         message_id = message.id
         message = MESSAGES.get(message_id, None)
-    elif isinstance(message, MessageReference):
-        channel_id = message.channel_id
-        message_id = message.message_id
-        message = MESSAGES.get(message_id, None)
     elif message is None:
         raise TypeError(
             '`message` was given as `None`. Make sure to call message create methods with non-empty content(s).'
@@ -1359,7 +1347,7 @@ def get_message_and_channel_id_and_message_id(message):
         snowflake_pair = maybe_snowflake_pair(message)
         if snowflake_pair is None:
             raise TypeError(
-                f'`message` can be `{Message.__name__}`, `{MessageRepr.__name__}`, `{MessageReference.__name__}`, '
+                f'`message` can be `{Message.__name__}`, `{MessageRepr.__name__}`, '
                 f'`tuple` of (`int`, `int`), got {message.__class__.__name__}; {message!r}.'
             )