+++ hata/discord/message/__init__.py
@@ -4,7 +4,6 @@ from .flags import *
 from .message import *
 from .message_application import *
 from .message_interaction import *
-from .message_reference import *
 from .message_repr import *
 from .preinstanced import *
 from .utils import *
@@ -16,8 +15,19 @@ __all__ = (
     *message.__all__,
     *message_application.__all__,
     *message_interaction.__all__,
-    *message_reference.__all__,
     *message_repr.__all__,
     *preinstanced.__all__,
     *utils.__all__,
 )
+
+
+from ...utils.module_deprecation import deprecated_import
+
+@deprecated_import
+class MessageReference:
+    """
+    A cross guild reference used as a ``Message``'s `.cross_reference` at crosspost messages.
+    
+    This type is Deprecated and will be removed in 2023 January.
+    """
+    __slots__ = ()