+++ changelog.md
@@ -1,9 +1,16 @@
 ## 1.3.4 *\[2022-08-??\]*
 
+#### Improvements
+
+- Add `Message._create_from_partial_data`.
+- `Message.cross_reference` now can be set only as a `Message` instance (or `None`).
+- Add `Message._create_from_partial_fields`.
+
 #### Renames, Deprecation & Removals
 
 - Rename `.is_bot` to `bot` to make it more intuitive.
 - Deprecate `.is_bot` everywhere
+- Deprecate `MessageReference`.
 
 ## 1.3.3 *\[2022-08-13\]*